+++ esp-wifi/src/preempt/preempt_riscv.rs
@@ -59,10 +59,8 @@ pub fn task_create(task: extern "C" fn()) {
         let task_stack_size = TASK_STACK_SIZE[i];
 
         // stack must be aligned by 16
-        let task_stack_ptr = addr_of!(TASK_STACK) as usize
-            + (task_stack_size as usize * i as usize)
-            + task_stack_size as usize
-            - 4;
+        let task_stack_ptr =
+            addr_of!(TASK_STACK) as usize + (task_stack_size * i) + task_stack_size - 4;
         let stack_ptr = task_stack_ptr - (task_stack_ptr % 0x10);
         CTX_TASKS[i].trap_frame.sp = stack_ptr;
     }
